ASoC: msm8x10-wcd: update the micbias count correctly
If headset is inserted during voice call, the micbias
count is updated. Now on removal of headset, the dapm
widget for force disable of micbias is not called. Due
to this when the headset is inserted again, the count
not updated for headset as force enable dapm widget
will not be called. Make sure to update the count only
in the micbias widget callback function.

CRs-Fixed: 735818
Change-Id: I9e666b19c48a942704471d3fec2d5a4939c973ea
Signed-off-by: Simmi Pateriya <simmip@codeaurora.org>